+++ src/twg2/parser/codeParser/KeywordUtil.java
@@ -20,6 +20,9 @@ public interface KeywordUtil<T_KEYWORD extends AccessModifier> {
 	 */
 	public T_KEYWORD tryToKeyword(String str);
 
+	/** Check if a string is an inheritance keyword (i.e. ':' in C# or 'extends' and 'implements' in Java and TypeScript */
+	public boolean isInheritanceKeyword(String str);
+
 	/** Check if a string is a keyword */
 	public boolean isKeyword(String str);