+++ test/twg2/parser/codeParser/test/CsParseFilesTest.java
@@ -3,22 +3,18 @@ package twg2.parser.codeParser.test;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashSet;
 import java.util.List;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
-import java.util.stream.Stream;
-
-import lombok.val;
 
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runners.Parameterized.Parameter;
 
 import twg2.ast.interm.classes.ClassAst;
-import twg2.collections.builder.ListUtil;
+import twg2.ast.interm.type.TypeSig.TypeSigResolved;
 import twg2.io.files.FileFormatException;
 import twg2.io.files.FileReadUtil;
 import twg2.parser.codeParser.csharp.CsBlock;
@@ -26,8 +22,8 @@ import twg2.parser.codeParser.tools.NameUtil;
 import twg2.parser.language.CodeLanguage;
 import twg2.parser.main.ParserMisc;
 import twg2.parser.project.ProjectClassSet;
+import twg2.parser.workflow.CodeFileParsed;
 import twg2.parser.workflow.CodeFileSrc;
-import checks.CheckCollections;
 
 /**
  * @author TeamworkGuy2
@@ -40,8 +36,6 @@ public class CsParseFilesTest {
 
 	@Parameter
 	private ProjectClassSet.Simple<CodeFileSrc<CodeLanguage>, CsBlock> projFiles;
-	@Parameter
-	private List<ClassAst.ResolvedImpl<CsBlock>> resClasses;
 
 
 	public CsParseFilesTest() throws IOException, FileFormatException {
@@ -55,50 +49,44 @@ public class CsParseFilesTest {
 
 		HashSet<List<String>> missingNamespaces = new HashSet<>();
 		ParserMisc.parseFileSet(Arrays.asList(trackSearchServiceFile, albumInfoFile, trackInfoFile), projFiles, executor, fileReader, null);
-		val resFileSet = ProjectClassSet.resolveClasses(projFiles, CsBlock.CLASS, missingNamespaces);
+		ProjectClassSet.Resolved<CodeFileSrc<CodeLanguage>, CsBlock> resFileSet = ProjectClassSet.resolveClasses(projFiles, CsBlock.CLASS, missingNamespaces);
 
-		val res = resFileSet.getCompilationUnitsStartWith(Arrays.asList(""));
+		List<CodeFileParsed.Resolved<CodeFileSrc<CodeLanguage>, CsBlock>> res = resFileSet.getCompilationUnitsStartWith(Arrays.asList(""));
 
 		// get a subset of all the parsed files
-		resClasses = new ArrayList<>();
-		for(val classInfo : res) {
-			resClasses.add(classInfo.getParsedClass());
+		for(CodeFileParsed.Resolved<CodeFileSrc<CodeLanguage>, CsBlock> classInfo : res) {
+			ClassAst.ResolvedImpl<CsBlock> classParsed = classInfo.getParsedClass();
+			String simpleName = classParsed.getSignature().getSimpleName();
+			if("ITrackSearchService".equals(simpleName)) {
+				trackSearchServiceDef = classParsed;
+			}
+			else if("TrackInfo".equals(simpleName)) {
+				trackInfoDef = classParsed;
+			}
+			else if("AlbumInfo".equals(simpleName)) {
+				albumInfoDef = classParsed;
+			}
+			else {
+				throw new IllegalStateException("unknown class '" + NameUtil.joinFqName(classParsed.getSignature().getFullName()) + "'");
+			}
 		}
-		resClasses.sort((c1, c2) -> NameUtil.joinFqName(c1.getSignature().getFullName()).compareTo(NameUtil.joinFqName(c2.getSignature().getFullName())));
-		trackSearchServiceDef = ensureOne(resClasses.stream().filter((t) -> "ITrackSearchService".equals(t.getSignature().getSimpleName())));
-		trackInfoDef = ensureOne(resClasses.stream().filter((t) -> "TrackInfo".equals(t.getSignature().getSimpleName())));
-		albumInfoDef = ensureOne(resClasses.stream().filter((t) -> "AlbumInfo".equals(t.getSignature().getSimpleName())));
 	}
 
 
 	@Test
 	public void checkResolvedNames() {
-		val classNames = ListUtil.map(resClasses, (ic) -> NameUtil.joinFqName(ic.getSignature().getFullName()));
-		CheckCollections.assertLooseEquals(Arrays.asList("ParserExamples.Services.ITrackSearchService", "ParserExamples.Models.AlbumInfo", "ParserExamples.Models.TrackInfo"), classNames);
-
 		// SearchResult<TrackInfo> Search(TrackSearchCriteria criteria)
-		val mthd1Ret = trackSearchServiceDef.getMethods().get(0).getReturnType();
+		TypeSigResolved mthd1Ret = trackSearchServiceDef.getMethods().get(0).getReturnType();
 		Assert.assertEquals("ParserExamples.Models.TrackInfo", NameUtil.joinFqName(mthd1Ret.getParams().get(0).getFullName()));
 
 		// SearchResult<IDictionary<AlbumInfo, IList<Track>>> GetAlbumTracks(string albumName)
-		val mthd2Ret = trackSearchServiceDef.getMethods().get(1).getReturnType();
+		TypeSigResolved mthd2Ret = trackSearchServiceDef.getMethods().get(1).getReturnType();
 
 		Assert.assertEquals("IDictionary", NameUtil.joinFqName(mthd2Ret.getParams().get(0).getFullName()));
 		Assert.assertEquals("ParserExamples.Models.AlbumInfo", NameUtil.joinFqName(mthd2Ret.getParams().get(0).getParams().get(0).getFullName()));
 		Assert.assertEquals("IList", NameUtil.joinFqName(mthd2Ret.getParams().get(0).getParams().get(1).getFullName()));
 		Assert.assertEquals("ParserExamples.Models.TrackInfo", NameUtil.joinFqName(mthd2Ret.getParams().get(0).getParams().get(1).getParams().get(0).getFullName()));
 
-	}
-
-
-	private static <R> R ensureOne(Stream<R> stream) {
-		Object[] objs = stream.toArray();
-		if(objs.length != 1) {
-			throw new IllegalArgumentException("Expected stream of 1 object, found " + objs.length + ": " + Arrays.toString(objs));
-		}
-		@SuppressWarnings("unchecked")
-		val res = (R)objs[0];
-		return res;
 	}
 
 }